merge_vertices accelerated by a KDTree
